Merged PR 6917: Add paper_index as the unique identifier
Add `paper_index` to the `Person` model.

This is the unique identifier for each response, written on each physical survey. It's generated as a counter that resets at the start of every time quarter. This generation may change after client feedback, but this PR sets the necessary infrastructure to use it generally.

`Paper_index` is now used as the URL route, and generally the unique identifier for each response across the application.

Includes:

- Add `paper_index` model field
- Update views to use `paper_index` over uuid
- Update urls to use `paper_index` over `UUID`
- Update templates to use `paper_index`
- Migrations for initial `paper_index`

Resolves: #65888

Related work items: #65888