Merged PR 6965: #65890 Add conditional forms on user permissions
Based on user feedback in the demo, depending on the users permission level, they will be able to "code" forms on initial entry, or not.
So have added the conditional on building the form to implement this.
This also has the effect of allowing users to see coded forms/data, and recode if necessary (65890)

Includes:
- Update permissions required explicitly to `Person` model
- Migrations for this change
- Add coding layout on user permissions in `forms.py`
- Update permission conditionals in templates
- Refactor all response forms to subclass `PersonForm`, simplifying forms.

Resolves #65890

Related work items: #65890